CI: revised clang-tidy list of checks

+++ src/block_template.cpp
@@ -865,7 +865,7 @@ void BlockTemplate::select_mempool_transactions(const Mempool& mempool)
 		k += r / 34359738368ULL;
 	}
 
-	const size_t max_transactions = (MAX_BLOCK_SIZE > k) ? ((MAX_BLOCK_SIZE - k) / HASH_SIZE) : 0;
+	const uint32_t max_transactions = static_cast<uint32_t>((MAX_BLOCK_SIZE > k) ? ((MAX_BLOCK_SIZE - k) / HASH_SIZE) : 0);
 	LOGINFO(6, max_transactions << " transactions can be taken with current block size limit");
 
 	if (max_transactions == 0) {
